filterInputAbstractの不要なstyledを統合
