+++ src/components/FilterInputAbstract/styled.tsx
@@ -1,42 +1,46 @@
-import styled from "styled-components";
+import styled, { css } from "styled-components";
 import { colors } from "../../styles";
 import { ContextMenu2CheckItem } from "../ContextMenu2/ContextMenu2CheckItem";
+import { FilterSize, FILTER_SIZES } from "./types";
 
-//
-// -----------------------------------------------------------------------------
-
-export const FilterTag = styled.span`
-  isolation: isolate;
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  width: fit-content;
-  padding: 1px 4px 1px 6px;
+// フィルターのベーススタイル
+export const filterBaseStyle = css`
   border: 1px solid ${colors.basic[400]};
-  border-radius: 2px;
-  /* UI/Text 12 */
-  font-weight: 400;
-  font-size: 12px;
-  line-height: 14px;
-  word-break: break-all;
-  color: ${colors.basic[900]};
   background-color: #fff;
 `;
 
-export const FilterTagButton = styled.button`
-  flex-shrink: 0;
-  height: 18px;
-  aspect-ratio: 1;
-  padding: 0;
-  border: 0;
-  zoom: ${16 / 18};
-  color: ${colors.basic[900]};
-  background-color: transparent;
-  cursor: pointer;
+// サイズに応じたスタイルを生成するヘルパー関数
+export const getFilterSizeStyle = (size: FilterSize) => css`
+  height: ${FILTER_SIZES[size].height};
+  border-radius: ${FILTER_SIZES[size].borderRadius};
 `;
 
-//
-// -----------------------------------------------------------------------------
+// トリガーのアイコンサイズスタイルを生成するヘルパー関数
+export const getTriggerIconSizeStyle = (size: FilterSize) => css`
+  /* 左側のアイコン */
+  span:first-child {
+    width: ${FILTER_SIZES[size].iconSize};
+    height: ${FILTER_SIZES[size].iconSize};
+    svg {
+      width: ${FILTER_SIZES[size].iconSize};
+      height: ${FILTER_SIZES[size].iconSize};
+    }
+  }
+  /* 右側の矢印アイコン */
+  span:last-child {
+    width: ${FILTER_SIZES[size].arrowIconSize};
+    height: ${FILTER_SIZES[size].arrowIconSize};
+    svg {
+      width: ${FILTER_SIZES[size].arrowIconSize};
+      height: ${FILTER_SIZES[size].arrowIconSize};
+    }
+  }
+  padding: ${FILTER_SIZES[size].padding};
+`;
+
+export type FilterTagProps = {
+  $size: FilterSize;
+};
 
 export const FilterInputAbstract = styled.div`
   position: relative;
@@ -44,24 +48,19 @@ export const FilterInputAbstract = styled.div`
   grid-template-columns: auto 1fr;
   align-items: center;
   gap: 0;
-  border-radius: 4px;
-  border: 1px solid ${colors.basic[400]};
-  background-color: #fff;
+  ${filterBaseStyle}
   overflow: hidden;
 
   &[data-size="small"] {
-    height: 28px;
-    border-radius: 4px;
+    ${getFilterSizeStyle("small")}
   }
 
   &[data-size="medium"] {
-    height: 32px;
-    border-radius: 6px;
+    ${getFilterSizeStyle("medium")}
   }
 
   &[data-size="large"] {
-    height: 40px;
-    border-radius: 6px;
+    ${getFilterSizeStyle("large")}
   }
 
   &[data-small="true"] {
@@ -79,7 +78,6 @@ export const DropDownTrigger = styled.button`
   column-gap: 2px;
   align-items: center;
   height: 100%;
-  padding: 0 2px 0 6px;
   border: 0;
   border-right: 1px solid ${colors.basic[400]};
   outline-offset: -1px;
@@ -89,71 +87,106 @@ export const DropDownTrigger = styled.button`
 
   /* サイズバリエーション */
   &:where(${FilterInputAbstract}[data-size="small"] *) {
-    /* 左側のアイコン */
-    span:first-child {
-      width: 20px;
-      height: 20px;
-      svg {
-        width: 20px;
-        height: 20px;
-      }
-    }
-    /* 右側の矢印アイコン */
-    span:last-child {
-      width: 16px;
-      height: 16px;
-      svg {
-        width: 16px;
-        height: 16px;
-      }
-    }
+    ${getTriggerIconSizeStyle("small")}
   }
 
   &:where(${FilterInputAbstract}[data-size="medium"] *) {
-    /* 左側のアイコン */
-    span:first-child {
-      width: 22px;
-      height: 22px;
-      svg {
-        width: 22px;
-        height: 22px;
-      }
-    }
-    /* 右側の矢印アイコン */
-    span:last-child {
-      width: 18px;
-      height: 18px;
-      svg {
-        width: 18px;
-        height: 18px;
-      }
-    }
+    ${getTriggerIconSizeStyle("medium")}
   }
 
   &:where(${FilterInputAbstract}[data-size="large"] *) {
-    padding: 0 4px 0 8px;
-    /* 左側のアイコン */
-    span:first-child {
-      width: 24px;
-      height: 24px;
-      svg {
-        width: 24px;
-        height: 24px;
-      }
-    }
-    /* 右側の矢印アイコン */
-    span:last-child {
-      width: 20px;
-      height: 20px;
-      svg {
-        width: 20px;
-        height: 20px;
-      }
-    }
+    ${getTriggerIconSizeStyle("large")}
   }
 
   &:where(${FilterInputAbstract.toString()}[data-small="true"] *) {
     display: none;
+  }
+`;
+
+export const FilterTag = styled.span<FilterTagProps>`
+  isolation: isolate;
+  display: flex;
+  align-items: center;
+  gap: 4px;
+  width: fit-content;
+  padding: ${({ $size }) => {
+    switch ($size) {
+      case "small":
+        return "2px 4px 2px 6px";
+      case "medium":
+        return "2px 5px 2px 7px";
+      case "large":
+        return "3px 6px 3px 8px";
+    }
+  }};
+  border: 1px solid ${colors.basic[400]};
+  border-radius: 2px;
+  font-weight: 400;
+  font-size: ${({ $size }) => {
+    switch ($size) {
+      case "small":
+        return "11px";
+      case "medium":
+        return "12px";
+      case "large":
+        return "13px";
+    }
+  }};
+  line-height: 14px;
+  word-break: break-all;
+  color: ${colors.basic[900]};
+  background-color: #fff;
+`;
+
+export const FilterTagButton = styled.button<FilterTagProps>`
+  flex-shrink: 0;
+  height: ${({ $size }) => {
+    switch ($size) {
+      case "small":
+        return "14px";
+      case "medium":
+        return "16px";
+      case "large":
+        return "18px";
+    }
+  }};
+  aspect-ratio: 1;
+  padding: 0;
+  border: 0;
+  color: ${colors.basic[900]};
+  background-color: transparent;
+  cursor: pointer;
+
+  /* アイコンのサイズ調整 */
+  svg {
+    width: ${({ $size }) => {
+    switch ($size) {
+      case "small":
+        return "14px";
+      case "medium":
+        return "16px";
+      case "large":
+        return "18px";
+    }
+  }};
+    height: ${({ $size }) => {
+    switch ($size) {
+      case "small":
+        return "14px";
+      case "medium":
+        return "16px";
+      case "large":
+        return "18px";
+    }
+  }};
+  }
+
+  span {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+    height: 100%;
   }
 `;
